Fix issues with Type::stream.
This patch moves Type::stream to rec_stream.cpp and calls Type::stream
in Def::stream.
